--- src/App.scss
@@ -1,3 +1,0 @@
-iframe {
-  display: none;
-}
